refactored: author dtos is moved to author package
